Move CookieController out of WebApi and into Website project.

+++ Mocktails.DAL/Model/MocktailQuantity.cs
@@ -22,7 +22,7 @@ public class MocktailQuantity
 
     public MocktailQuantity() { }
 
-    public int GetTotalPrice()
+    public decimal GetTotalPrice()
     {
         return Price * Quantity;
     }
